[FIX] Website: Only one of the multiple fields are visible when more than one of the multiple checkboxes are selected
The issue was found inside /odoo/addons/website/static/src/snippets/s_website_form/000.js, the _buildVisibilityFunction() that was calculating the visibility of each field having a visibility condition.
FormData.get() was supposed to fetch the dependencyName when a checkbox was being selected, but when multiple checkboxes were being selected it was returning only one value, the first value. Only the first field was being visible whenever the multiple checkboxes were selected.

Description of the changes made:
Replaced FormData.get(dependencyName) with FormData.getAll(dependencyName) and stored in selected variable
Added an if condition  if(selected.includes(visibilityCondition))
returning the visibilityCondition if visibilityCondition is present in the values returned by FormData.getAll(dependencyName) resulting in all the fields having the visibility dependency to be visible instead of only the first one else returning 'undefined' so that none of the fields are visible as they have not been selected.